move all clustering code to cluster folder
