+++ spateo/tools/__init__.py
@@ -1,10 +1,11 @@
+from spateo.tools.cluster.find_clusters import scc, spagcn_pyg
+
 from .cluster import *
 from .cluster_degs import (
     find_all_cluster_degs,
     find_cluster_degs,
     find_spatial_cluster_degs,
 )
-from .find_clusters import find_cluster_scc, find_cluster_spagcn
 
 # from .image import add_image_layer
 # from .interpolation_utils import *
