+++ spateo/tools/cluster/__init__.py
@@ -1,5 +1,5 @@
-from .cluster_dynamo import cluster_dyn, spatial_adj_dyn
-from .cluster_spagcn import cluster_spagcn
+from .cluster_spagcn import spagcn_vanilla
+from .find_clusters import scc, spagcn_pyg
 from .utils import (
     compute_pca_components,
     ecp_silhouette,
